日报：未涉及的项目和涉及项目还有综合项目进行分类展示，更明显。 日报：未填写后提交成功的bug处理。 日报：按照工作日历来展示工作日和休息日。 日报：时长问题，后面的时长加个自定义输入框。 员工日报：展示人员上次填写日报天数 项目查看：项目详情-金额全部隐藏掉。 项目管理：项目id选择区域时加地区首字母检索 项目管理：项目id选择完项目类型时后面展示类型备注 项目管理：删除项目时判断是否有工时，有则提示无法删除 综合项目管理：开发综合项目模块，只展示项目类型为综合（ZH）类的项目，进行日常其他项目的维护和管理，拥有项目经理，组长和主管权限人可以查看 综合项目日报：开发综合项目模块，展示综合项目日报填写情况，内容和项目管理内日报类似，拥有项目经理,组长和主管权限人可以查看 市场客户管理：新建市场客户，方便市场人员维护数据和填写拜访记录。

+++ pm-resource-server/src/graphql/resolvers/projs.ts
@@ -13,8 +13,17 @@ export default {
       .sort({ createDate: -1 })
       .map(dbid2id).toArray()
     },
+    superProjs: (_: any, __: any, context: AuthContext) => { 
+      let filter = { _id: { $not: /-ZH-/ } }
+      if (__.isArchive === true || __.isArchive === false) { filter['isArchive'] = __.isArchive }
+      return Project
+      .find(filter)
+      .sort({ createDate: -1 })
+      .map(dbid2id).toArray()
+    },
     iLeadProjs: (_: any, __: any, context: AuthContext) => Project
       .find({
+        _id: { $not: /-ZH-/ },
         isArchive: __.isArchive ? __.isArchive : false,
         $or: [ { leader: context.user!.id, }, { salesLeader: context.user!.id } ],
       })
@@ -24,7 +33,14 @@ export default {
       .find({
         isArchive: false,
         $or: [ { leader: context.user!.id, }, { salesLeader: context.user!.id, }, { participants: { $elemMatch: { $eq: context.user!.id } } } ],
-        _id: { $regex: '(?:[^-]+-){1}'+__.projType+'' },
+        _id: { $regex: `^[0-9A-Za-z]*-[0-9A-Za-z]*-${__.projType}+-[0-9A-Za-z]*-[0-9A-Za-z]*$` },
+      })
+      .sort({ createDate: -1 })
+      .map(dbid2id).toArray(),
+    filterProjsByType: (_: any, __: any, context: AuthContext) => Project
+      .find({
+        isArchive: false,
+        _id: { $regex: `^[0-9A-Za-z]*-[0-9A-Za-z]*-${__.projType}+-[0-9A-Za-z]*-[0-9A-Za-z]*$` },
       })
       .sort({ createDate: -1 })
       .map(dbid2id).toArray(),
@@ -81,6 +97,10 @@ export default {
       return Project.updateOne({ _id: args.id }, { $set: { isArchive: true, archiveTime, archivePerson: context.user!.id } }, { upsert: true }).then((res) => args.id || res.upsertedId._id)
     },
     deleteProject: async (_: any, args: any, context: AuthContext) => {
+      let deletProj = await Project.findOne({ _id: args.id })
+      if (deletProj?.timeConsuming) {
+        return new Error(`项目已存在工时，无法删除！`)
+      }
       // 清除合同的绑定关系
       await ProjectAgreement.deleteMany({ _id: args.id })
       return Project.deleteOne({ _id: args.id }).then(() => args.id)
